Lots of changes ;-)
Changed OSX detection
Removed LotsOfButtonsTable: Performance creating that many widgets on-the-fly was ok for small parse states, but large ones killed all productivity of using the GUI for testing, moved to a modified version of GridView
Added GirdView (from Openvista); fixed bugs and added cell event
Added the Mono.TextEditor Syntax modes to enable the basics of highlighting. Using this method could allow full dynamic highlighting and Intellisense if the need was there. Might be a fun task but the return on time might be low if only used within this tool.
Misc code cleanup and minor refactoring (more need, but not required for current feature set)